Adjust `Salary.getSalaryActual` to use float division for accurate calculation and add test for fractional salary deduction.

+++ src/main/java/com/example/demo/Salary.java
@@ -3,6 +3,6 @@ package com.example.demo;
 public record Salary(int value) {
 
     public int getSalaryActual(long monthDays, int leaveDays) {
-        return (int) (value - value * leaveDays / monthDays);
+        return (int) (value - (float) value * leaveDays / monthDays);
     }
 }
